refactor: arranged file tree
