+++ frontend/components/gmm-home/gmm-num-component-selector/hooks/use-num-components.tsx
@@ -1,0 +1,67 @@
+import { range } from "lodash";
+import { useRouter } from "next/router";
+import { z } from "zod";
+import { apiClient } from "~/services/api-client";
+import { responseGetGMMJobsItems } from "~/services/route/gmm";
+
+export type Job = z.infer<typeof responseGetGMMJobsItems>;
+
+export const useNumComponents = (item: Job) => {
+  const { push } = useRouter();
+  const { status, uuid, params } = item;
+
+  if (status === "failure" || status === "pending") {
+    return {
+      value: NaN,
+      optimalValue: NaN,
+      range: [],
+      onSelect: () => {},
+      onSubmit: async () => {
+        throw new Error("Invalid job submitted");
+      },
+    };
+  }
+
+  const numComponents = range(
+    params.minimum_n_components,
+    params.maximum_n_components + 1,
+    params.step_size
+  );
+
+  const onSelect = (e: React.ChangeEvent<HTMLSelectElement>) => {
+    try {
+      const n = parseInt(e.currentTarget.value);
+      if (!numComponents.includes(n)) {
+        throw new Error("Invalid number of components");
+      }
+      push(`?experiment=${item.uuid}&n_components=${n}`, undefined, {
+        scroll: false,
+      });
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  const onSubmit = async (name: string) => {
+    if (item.status !== "success") {
+      return;
+    }
+    try {
+      await apiClient.publishGMMJobs({
+        name: name,
+        uuid: uuid,
+        n_components: item.gmm.current_n_components,
+      });
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  return {
+    value: item.gmm.current_n_components,
+    optimalValue: item.gmm.current_n_components,
+    range: numComponents,
+    onSelect,
+    onSubmit,
+  };
+};
