--- frontend/components/trainer-home/job-params/index.tsx
@@ -1,2 +1,0 @@
-export { ParentJobParams } from "./parent-job-params";
-export { ChildJobParams } from "./child-job-params";
