+++ frontend/components/trainer-home/training-params/index.tsx
@@ -1,11 +1,28 @@
 import React from "react";
 import { CustomDataGrid } from "~/components/common/custom-datagrid";
-import { ParamsType, useTrainingParams } from "./hooks/use-training-params";
+import { z } from "zod";
+import { responseGetItem } from "~/services/route/train";
+
+export type ParamsType = z.infer<typeof responseGetItem>["params_training"];
 
 export const TrainingParams: React.FC<{
   value: ParamsType;
-}> = (props) => {
-  const { dataSource, columns, gridStyle } = useTrainingParams(props.value);
+}> = ({ value }) => {
+  const dataSource = Object.entries(value).map(([k, v], index) => ({
+    id: index,
+    item: k,
+    value: String(v),
+  }));
+
+  const columns = [
+    { name: "id", type: "number", header: "ID", defaultVisible: false },
+    { name: "item", header: "Item", defaultFlex: 1 },
+    { name: "value", header: "Value" },
+  ];
+
+  const gridStyle = {
+    height: 300,
+  };
 
   return (
     <CustomDataGrid
